Split a lot of code from win32_edge.hh
