+++ core/include/webview/detail/backends/win32_edge.hh
@@ -40,17 +40,19 @@
 #include "../engine_base.hh"
 #include "../native_library.hh"
 #include "../platform/windows/com_init_wrapper.hh"
+#include "../platform/windows/dpi.hh"
 #include "../platform/windows/reg_key.hh"
+#include "../platform/windows/theme.hh"
+#include "../platform/windows/version.hh"
 #include "../user_script.hh"
 #include "../utility/string.hh"
+#include "webview2/loader.hh"
 
-#include <array>
 #include <atomic>
 #include <functional>
 #include <list>
 #include <memory>
 #include <utility>
-#include <vector>
 
 #ifndef WIN32_LEAN_AND_MEAN
 #define WIN32_LEAN_AND_MEAN
@@ -61,10 +63,7 @@
 #include <stdlib.h>
 #include <windows.h>
 
-#include "WebView2.h"
-
 #ifdef _MSC_VER
-#pragma comment(lib, "advapi32.lib")
 #pragma comment(lib, "ole32.lib")
 #pragma comment(lib, "shell32.lib")
 #pragma comment(lib, "shlwapi.lib")
@@ -76,614 +75,6 @@ namespace webview {
 namespace detail {
 
 using msg_cb_t = std::function<void(const std::string)>;
-
-// Parses a version string with 1-4 integral components, e.g. "1.2.3.4".
-// Missing or invalid components default to 0, and excess components are ignored.
-template <typename T>
-std::array<unsigned int, 4>
-parse_version(const std::basic_string<T> &version) noexcept {
-  auto parse_component = [](auto sb, auto se) -> unsigned int {
-    try {
-      auto n = std::stol(std::basic_string<T>(sb, se));
-      return n < 0 ? 0 : n;
-    } catch (std::exception &) {
-      return 0;
-    }
-  };
-  auto end = version.end();
-  auto sb = version.begin(); // subrange begin
-  auto se = sb;              // subrange end
-  unsigned int ci = 0;       // component index
-  std::array<unsigned int, 4> components{};
-  while (sb != end && se != end && ci < components.size()) {
-    if (*se == static_cast<T>('.')) {
-      components[ci++] = parse_component(sb, se);
-      sb = ++se;
-      continue;
-    }
-    ++se;
-  }
-  if (sb < se && ci < components.size()) {
-    components[ci] = parse_component(sb, se);
-  }
-  return components;
-}
-
-template <typename T, std::size_t Length>
-auto parse_version(const T (&version)[Length]) noexcept {
-  return parse_version(std::basic_string<T>(version, Length));
-}
-
-inline std::wstring
-get_file_version_string(const std::wstring &file_path) noexcept {
-  DWORD dummy_handle; // Unused
-  DWORD info_buffer_length =
-      GetFileVersionInfoSizeW(file_path.c_str(), &dummy_handle);
-  if (info_buffer_length == 0) {
-    return std::wstring();
-  }
-  std::vector<char> info_buffer;
-  info_buffer.reserve(info_buffer_length);
-  if (!GetFileVersionInfoW(file_path.c_str(), 0, info_buffer_length,
-                           info_buffer.data())) {
-    return std::wstring();
-  }
-  auto sub_block = L"\\StringFileInfo\\040904B0\\ProductVersion";
-  LPWSTR version = nullptr;
-  unsigned int version_length = 0;
-  if (!VerQueryValueW(info_buffer.data(), sub_block,
-                      reinterpret_cast<LPVOID *>(&version), &version_length)) {
-    return std::wstring();
-  }
-  if (!version || version_length == 0) {
-    return std::wstring();
-  }
-  return std::wstring(version, version_length);
-}
-
-namespace ntdll_symbols {
-using RtlGetVersion_t =
-    unsigned int /*NTSTATUS*/ (WINAPI *)(RTL_OSVERSIONINFOW *);
-
-constexpr auto RtlGetVersion = library_symbol<RtlGetVersion_t>("RtlGetVersion");
-} // namespace ntdll_symbols
-
-namespace user32_symbols {
-using DPI_AWARENESS_CONTEXT = HANDLE;
-using SetProcessDpiAwarenessContext_t = BOOL(WINAPI *)(DPI_AWARENESS_CONTEXT);
-using SetProcessDPIAware_t = BOOL(WINAPI *)();
-using GetDpiForWindow_t = UINT(WINAPI *)(HWND);
-using EnableNonClientDpiScaling_t = BOOL(WINAPI *)(HWND);
-using AdjustWindowRectExForDpi_t = BOOL(WINAPI *)(LPRECT, DWORD, BOOL, DWORD,
-                                                  UINT);
-using GetWindowDpiAwarenessContext_t = DPI_AWARENESS_CONTEXT(WINAPI *)(HWND);
-using AreDpiAwarenessContextsEqual_t = BOOL(WINAPI *)(DPI_AWARENESS_CONTEXT,
-                                                      DPI_AWARENESS_CONTEXT);
-
-// Use intptr_t as the underlying type because we need to
-// reinterpret_cast<DPI_AWARENESS_CONTEXT> which is a pointer.
-// Available since Windows 10, version 1607
-enum class dpi_awareness : intptr_t {
-  per_monitor_v2_aware = -4, // Available since Windows 10, version 1703
-  per_monitor_aware = -3
-};
-
-constexpr auto SetProcessDpiAwarenessContext =
-    library_symbol<SetProcessDpiAwarenessContext_t>(
-        "SetProcessDpiAwarenessContext");
-constexpr auto SetProcessDPIAware =
-    library_symbol<SetProcessDPIAware_t>("SetProcessDPIAware");
-constexpr auto GetDpiForWindow =
-    library_symbol<GetDpiForWindow_t>("GetDpiForWindow");
-constexpr auto EnableNonClientDpiScaling =
-    library_symbol<EnableNonClientDpiScaling_t>("EnableNonClientDpiScaling");
-constexpr auto AdjustWindowRectExForDpi =
-    library_symbol<AdjustWindowRectExForDpi_t>("AdjustWindowRectExForDpi");
-constexpr auto GetWindowDpiAwarenessContext =
-    library_symbol<GetWindowDpiAwarenessContext_t>(
-        "GetWindowDpiAwarenessContext");
-constexpr auto AreDpiAwarenessContextsEqual =
-    library_symbol<AreDpiAwarenessContextsEqual_t>(
-        "AreDpiAwarenessContextsEqual");
-} // namespace user32_symbols
-
-namespace dwmapi_symbols {
-typedef enum {
-  // This undocumented value is used instead of DWMWA_USE_IMMERSIVE_DARK_MODE
-  // on Windows 10 older than build 19041 (2004/20H1).
-  DWMWA_USE_IMMERSIVE_DARK_MODE_BEFORE_V10_0_19041 = 19,
-  // Documented as being supported since Windows 11 build 22000 (21H2) but it
-  // works since Windows 10 build 19041 (2004/20H1).
-  DWMWA_USE_IMMERSIVE_DARK_MODE = 20
-} DWMWINDOWATTRIBUTE;
-using DwmSetWindowAttribute_t = HRESULT(WINAPI *)(HWND, DWORD, LPCVOID, DWORD);
-
-constexpr auto DwmSetWindowAttribute =
-    library_symbol<DwmSetWindowAttribute_t>("DwmSetWindowAttribute");
-} // namespace dwmapi_symbols
-
-namespace shcore_symbols {
-typedef enum { PROCESS_PER_MONITOR_DPI_AWARE = 2 } PROCESS_DPI_AWARENESS;
-using SetProcessDpiAwareness_t = HRESULT(WINAPI *)(PROCESS_DPI_AWARENESS);
-
-constexpr auto SetProcessDpiAwareness =
-    library_symbol<SetProcessDpiAwareness_t>("SetProcessDpiAwareness");
-} // namespace shcore_symbols
-
-// Compare the specified version against the OS version.
-// Returns less than 0 if the OS version is less.
-// Returns 0 if the versions are equal.
-// Returns greater than 0 if the specified version is greater.
-inline int compare_os_version(unsigned int major, unsigned int minor,
-                              unsigned int build) {
-  // Use RtlGetVersion both to bypass potential issues related to
-  // VerifyVersionInfo and manifests, and because both GetVersion and
-  // GetVersionEx are deprecated.
-  auto ntdll = native_library(L"ntdll.dll");
-  if (auto fn = ntdll.get(ntdll_symbols::RtlGetVersion)) {
-    RTL_OSVERSIONINFOW vi{};
-    vi.dwOSVersionInfoSize = sizeof(vi);
-    if (fn(&vi) != 0) {
-      return false;
-    }
-    if (vi.dwMajorVersion == major) {
-      if (vi.dwMinorVersion == minor) {
-        return static_cast<int>(vi.dwBuildNumber) - static_cast<int>(build);
-      }
-      return static_cast<int>(vi.dwMinorVersion) - static_cast<int>(minor);
-    }
-    return static_cast<int>(vi.dwMajorVersion) - static_cast<int>(major);
-  }
-  return false;
-}
-
-inline bool is_per_monitor_v2_awareness_available() {
-  // Windows 10, version 1703
-  return compare_os_version(10, 0, 15063) >= 0;
-}
-
-inline bool enable_dpi_awareness() {
-  auto user32 = native_library(L"user32.dll");
-  if (auto fn = user32.get(user32_symbols::SetProcessDpiAwarenessContext)) {
-    auto dpi_awareness =
-        reinterpret_cast<user32_symbols::DPI_AWARENESS_CONTEXT>(
-            is_per_monitor_v2_awareness_available()
-                ? user32_symbols::dpi_awareness::per_monitor_v2_aware
-                : user32_symbols::dpi_awareness::per_monitor_aware);
-    if (fn(dpi_awareness)) {
-      return true;
-    }
-    return GetLastError() == ERROR_ACCESS_DENIED;
-  }
-  if (auto shcore = native_library(L"shcore.dll")) {
-    if (auto fn = shcore.get(shcore_symbols::SetProcessDpiAwareness)) {
-      auto result = fn(shcore_symbols::PROCESS_PER_MONITOR_DPI_AWARE);
-      return result == S_OK || result == E_ACCESSDENIED;
-    }
-  }
-  if (auto fn = user32.get(user32_symbols::SetProcessDPIAware)) {
-    return !!fn();
-  }
-  return true;
-}
-
-inline bool enable_non_client_dpi_scaling_if_needed(HWND window) {
-  auto user32 = native_library(L"user32.dll");
-  auto get_ctx_fn = user32.get(user32_symbols::GetWindowDpiAwarenessContext);
-  if (!get_ctx_fn) {
-    return true;
-  }
-  auto awareness = get_ctx_fn(window);
-  if (!awareness) {
-    return false;
-  }
-  auto ctx_equal_fn = user32.get(user32_symbols::AreDpiAwarenessContextsEqual);
-  if (!ctx_equal_fn) {
-    return true;
-  }
-  // EnableNonClientDpiScaling is only needed with per monitor v1 awareness.
-  auto per_monitor = reinterpret_cast<user32_symbols::DPI_AWARENESS_CONTEXT>(
-      user32_symbols::dpi_awareness::per_monitor_aware);
-  if (!ctx_equal_fn(awareness, per_monitor)) {
-    return true;
-  }
-  auto enable_fn = user32.get(user32_symbols::EnableNonClientDpiScaling);
-  if (!enable_fn) {
-    return true;
-  }
-  return !!enable_fn(window);
-}
-
-constexpr int get_default_window_dpi() {
-  constexpr const int default_dpi = 96; // USER_DEFAULT_SCREEN_DPI
-  return default_dpi;
-}
-
-inline int get_window_dpi(HWND window) {
-  auto user32 = native_library(L"user32.dll");
-  if (auto fn = user32.get(user32_symbols::GetDpiForWindow)) {
-    auto dpi = static_cast<int>(fn(window));
-    return dpi;
-  }
-  return get_default_window_dpi();
-}
-
-constexpr int scale_value_for_dpi(int value, int from_dpi, int to_dpi) {
-  return (value * to_dpi) / from_dpi;
-}
-
-constexpr SIZE scale_size(int width, int height, int from_dpi, int to_dpi) {
-  auto scaled_width = scale_value_for_dpi(width, from_dpi, to_dpi);
-  auto scaled_height = scale_value_for_dpi(height, from_dpi, to_dpi);
-  return {scaled_width, scaled_height};
-}
-
-inline SIZE make_window_frame_size(HWND window, int width, int height,
-                                   int dpi) {
-  auto style = GetWindowLong(window, GWL_STYLE);
-  RECT r{0, 0, width, height};
-  auto user32 = native_library(L"user32.dll");
-  if (auto fn = user32.get(user32_symbols::AdjustWindowRectExForDpi)) {
-    fn(&r, style, FALSE, 0, static_cast<UINT>(dpi));
-  } else {
-    AdjustWindowRect(&r, style, 0);
-  }
-  auto frame_width = r.right - r.left;
-  auto frame_height = r.bottom - r.top;
-  return {frame_width, frame_height};
-}
-
-inline bool is_dark_theme_enabled() {
-  constexpr auto *sub_key =
-      L"SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize";
-  reg_key key(HKEY_CURRENT_USER, sub_key, 0, KEY_READ);
-  if (!key.is_open()) {
-    // Default is light theme
-    return false;
-  }
-  return key.query_uint(L"AppsUseLightTheme", 1) == 0;
-}
-
-inline void apply_window_theme(HWND window) {
-  auto dark_theme_enabled = is_dark_theme_enabled();
-
-  // Use "immersive dark mode" on systems that support it.
-  // Changes the color of the window's title bar (light or dark).
-  BOOL use_dark_mode{dark_theme_enabled ? TRUE : FALSE};
-  static native_library dwmapi{L"dwmapi.dll"};
-  if (auto fn = dwmapi.get(dwmapi_symbols::DwmSetWindowAttribute)) {
-    // Try the modern, documented attribute before the older, undocumented one.
-    if (fn(window, dwmapi_symbols::DWMWA_USE_IMMERSIVE_DARK_MODE,
-           &use_dark_mode, sizeof(use_dark_mode)) != S_OK) {
-      fn(window,
-         dwmapi_symbols::DWMWA_USE_IMMERSIVE_DARK_MODE_BEFORE_V10_0_19041,
-         &use_dark_mode, sizeof(use_dark_mode));
-    }
-  }
-}
-
-// Enable built-in WebView2Loader implementation by default.
-#ifndef WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL
-#define WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL 1
-#endif
-
-// Link WebView2Loader.dll explicitly by default only if the built-in
-// implementation is enabled.
-#ifndef WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK
-#define WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL
-#endif
-
-// Explicit linking of WebView2Loader.dll should be used along with
-// the built-in implementation.
-#if WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL == 1 &&                                    \
-    WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK != 1
-#undef WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK
-#error Please set WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK=1.
-#endif
-
-#if WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL == 1
-// Gets the last component of a Windows native file path.
-// For example, if the path is "C:\a\b" then the result is "b".
-template <typename T>
-std::basic_string<T>
-get_last_native_path_component(const std::basic_string<T> &path) {
-  auto pos = path.find_last_of(static_cast<T>('\\'));
-  if (pos != std::basic_string<T>::npos) {
-    return path.substr(pos + 1);
-  }
-  return std::basic_string<T>();
-}
-#endif // WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL
-
-template <typename T> struct cast_info_t {
-  using type = T;
-  IID iid;
-};
-
-namespace mswebview2 {
-static constexpr IID
-    IID_ICoreWebView2CreateCoreWebView2ControllerCompletedHandler{
-        0x6C4819F3,
-        0xC9B7,
-        0x4260,
-        {0x81, 0x27, 0xC9, 0xF5, 0xBD, 0xE7, 0xF6, 0x8C}};
-static constexpr IID
-    IID_ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler{
-        0x4E8A3389,
-        0xC9D8,
-        0x4BD2,
-        {0xB6, 0xB5, 0x12, 0x4F, 0xEE, 0x6C, 0xC1, 0x4D}};
-static constexpr IID IID_ICoreWebView2PermissionRequestedEventHandler{
-    0x15E1C6A3,
-    0xC72A,
-    0x4DF3,
-    {0x91, 0xD7, 0xD0, 0x97, 0xFB, 0xEC, 0x6B, 0xFD}};
-static constexpr IID IID_ICoreWebView2WebMessageReceivedEventHandler{
-    0x57213F19,
-    0x00E6,
-    0x49FA,
-    {0x8E, 0x07, 0x89, 0x8E, 0xA0, 0x1E, 0xCB, 0xD2}};
-static constexpr IID
-    IID_ICoreWebView2AddScriptToExecuteOnDocumentCreatedCompletedHandler{
-        0xB99369F3,
-        0x9B11,
-        0x47B5,
-        {0xBC, 0x6F, 0x8E, 0x78, 0x95, 0xFC, 0xEA, 0x17}};
-
-#if WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL == 1
-enum class webview2_runtime_type { installed = 0, embedded = 1 };
-
-namespace webview2_symbols {
-using CreateWebViewEnvironmentWithOptionsInternal_t =
-    HRESULT(STDMETHODCALLTYPE *)(
-        bool, webview2_runtime_type, PCWSTR, IUnknown *,
-        ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler *);
-using DllCanUnloadNow_t = HRESULT(STDMETHODCALLTYPE *)();
-
-static constexpr auto CreateWebViewEnvironmentWithOptionsInternal =
-    library_symbol<CreateWebViewEnvironmentWithOptionsInternal_t>(
-        "CreateWebViewEnvironmentWithOptionsInternal");
-static constexpr auto DllCanUnloadNow =
-    library_symbol<DllCanUnloadNow_t>("DllCanUnloadNow");
-} // namespace webview2_symbols
-#endif // WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL
-
-#if WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK == 1
-namespace webview2_symbols {
-using CreateCoreWebView2EnvironmentWithOptions_t = HRESULT(STDMETHODCALLTYPE *)(
-    PCWSTR, PCWSTR, ICoreWebView2EnvironmentOptions *,
-    ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler *);
-using GetAvailableCoreWebView2BrowserVersionString_t =
-    HRESULT(STDMETHODCALLTYPE *)(PCWSTR, LPWSTR *);
-
-static constexpr auto CreateCoreWebView2EnvironmentWithOptions =
-    library_symbol<CreateCoreWebView2EnvironmentWithOptions_t>(
-        "CreateCoreWebView2EnvironmentWithOptions");
-static constexpr auto GetAvailableCoreWebView2BrowserVersionString =
-    library_symbol<GetAvailableCoreWebView2BrowserVersionString_t>(
-        "GetAvailableCoreWebView2BrowserVersionString");
-} // namespace webview2_symbols
-#endif // WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK
-
-class loader {
-public:
-  HRESULT create_environment_with_options(
-      PCWSTR browser_dir, PCWSTR user_data_dir,
-      ICoreWebView2EnvironmentOptions *env_options,
-      ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler
-          *created_handler) const {
-#if WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK == 1
-    if (m_lib.is_loaded()) {
-      if (auto fn = m_lib.get(
-              webview2_symbols::CreateCoreWebView2EnvironmentWithOptions)) {
-        return fn(browser_dir, user_data_dir, env_options, created_handler);
-      }
-    }
-#if WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL == 1
-    return create_environment_with_options_impl(browser_dir, user_data_dir,
-                                                env_options, created_handler);
-#else
-    return S_FALSE;
-#endif
-#else
-    return ::CreateCoreWebView2EnvironmentWithOptions(
-        browser_dir, user_data_dir, env_options, created_handler);
-#endif // WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK
-  }
-
-  HRESULT
-  get_available_browser_version_string(PCWSTR browser_dir,
-                                       LPWSTR *version) const {
-#if WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK == 1
-    if (m_lib.is_loaded()) {
-      if (auto fn = m_lib.get(
-              webview2_symbols::GetAvailableCoreWebView2BrowserVersionString)) {
-        return fn(browser_dir, version);
-      }
-    }
-#if WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL == 1
-    return get_available_browser_version_string_impl(browser_dir, version);
-#else
-    return S_FALSE;
-#endif
-#else
-    return ::GetAvailableCoreWebView2BrowserVersionString(browser_dir, version);
-#endif // WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK
-  }
-
-private:
-#if WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL == 1
-  struct client_info_t {
-    bool found = false;
-    std::wstring dll_path;
-    std::wstring version;
-    webview2_runtime_type runtime_type;
-  };
-
-  HRESULT create_environment_with_options_impl(
-      PCWSTR browser_dir, PCWSTR user_data_dir,
-      ICoreWebView2EnvironmentOptions *env_options,
-      ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler
-          *created_handler) const {
-    auto found_client = find_available_client(browser_dir);
-    if (!found_client.found) {
-      return -1;
-    }
-    auto client_dll = native_library(found_client.dll_path.c_str());
-    if (auto fn = client_dll.get(
-            webview2_symbols::CreateWebViewEnvironmentWithOptionsInternal)) {
-      return fn(true, found_client.runtime_type, user_data_dir, env_options,
-                created_handler);
-    }
-    if (auto fn = client_dll.get(webview2_symbols::DllCanUnloadNow)) {
-      if (!fn()) {
-        client_dll.detach();
-      }
-    }
-    return ERROR_SUCCESS;
-  }
-
-  HRESULT
-  get_available_browser_version_string_impl(PCWSTR browser_dir,
-                                            LPWSTR *version) const {
-    if (!version) {
-      return -1;
-    }
-    auto found_client = find_available_client(browser_dir);
-    if (!found_client.found) {
-      return -1;
-    }
-    auto info_length_bytes =
-        found_client.version.size() * sizeof(found_client.version[0]);
-    auto info = static_cast<LPWSTR>(CoTaskMemAlloc(info_length_bytes));
-    if (!info) {
-      return -1;
-    }
-    CopyMemory(info, found_client.version.c_str(), info_length_bytes);
-    *version = info;
-    return 0;
-  }
-
-  client_info_t find_available_client(PCWSTR browser_dir) const {
-    if (browser_dir) {
-      return find_embedded_client(api_version, browser_dir);
-    }
-    auto found_client =
-        find_installed_client(api_version, true, default_release_channel_guid);
-    if (!found_client.found) {
-      found_client = find_installed_client(api_version, false,
-                                           default_release_channel_guid);
-    }
-    return found_client;
-  }
-
-  std::wstring make_client_dll_path(const std::wstring &dir) const {
-    auto dll_path = dir;
-    if (!dll_path.empty()) {
-      auto last_char = dir[dir.size() - 1];
-      if (last_char != L'\\' && last_char != L'/') {
-        dll_path += L'\\';
-      }
-    }
-    dll_path += L"EBWebView\\";
-#if defined(_M_X64) || defined(__x86_64__)
-    dll_path += L"x64";
-#elif defined(_M_IX86) || defined(__i386__)
-    dll_path += L"x86";
-#elif defined(_M_ARM64) || defined(__aarch64__)
-    dll_path += L"arm64";
-#else
-#error WebView2 integration for this platform is not yet supported.
-#endif
-    dll_path += L"\\EmbeddedBrowserWebView.dll";
-    return dll_path;
-  }
-
-  client_info_t
-  find_installed_client(unsigned int min_api_version, bool system,
-                        const std::wstring &release_channel) const {
-    std::wstring sub_key = client_state_reg_sub_key;
-    sub_key += release_channel;
-    auto root_key = system ? HKEY_LOCAL_MACHINE : HKEY_CURRENT_USER;
-    reg_key key(root_key, sub_key, 0, KEY_READ | KEY_WOW64_32KEY);
-    if (!key.is_open()) {
-      return {};
-    }
-    auto ebwebview_value = key.query_string(L"EBWebView");
-
-    auto client_version_string =
-        get_last_native_path_component(ebwebview_value);
-    auto client_version = parse_version(client_version_string);
-    if (client_version[2] < min_api_version) {
-      // Our API version is greater than the runtime API version.
-      return {};
-    }
-
-    auto client_dll_path = make_client_dll_path(ebwebview_value);
-    return {true, client_dll_path, client_version_string,
-            webview2_runtime_type::installed};
-  }
-
-  client_info_t find_embedded_client(unsigned int min_api_version,
-                                     const std::wstring &dir) const {
-    auto client_dll_path = make_client_dll_path(dir);
-
-    auto client_version_string = get_file_version_string(client_dll_path);
-    auto client_version = parse_version(client_version_string);
-    if (client_version[2] < min_api_version) {
-      // Our API version is greater than the runtime API version.
-      return {};
-    }
-
-    return {true, client_dll_path, client_version_string,
-            webview2_runtime_type::embedded};
-  }
-
-  // The minimum WebView2 API version we need regardless of the SDK release
-  // actually used. The number comes from the SDK release version,
-  // e.g. 1.0.1150.38. To be safe the SDK should have a number that is greater
-  // than or equal to this number. The Edge browser webview client must
-  // have a number greater than or equal to this number.
-  static constexpr unsigned int api_version = 1150;
-
-  static constexpr auto client_state_reg_sub_key =
-      L"SOFTWARE\\Microsoft\\EdgeUpdate\\ClientState\\";
-
-  // GUID for the stable release channel.
-  static constexpr auto stable_release_guid =
-      L"{F3017226-FE2A-4295-8BDF-00C3A9A7E4C5}";
-
-  static constexpr auto default_release_channel_guid = stable_release_guid;
-#endif // WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL
-
-#if WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK == 1
-  native_library m_lib{L"WebView2Loader.dll"};
-#endif
-};
-
-namespace cast_info {
-static constexpr auto controller_completed =
-    cast_info_t<ICoreWebView2CreateCoreWebView2ControllerCompletedHandler>{
-        IID_ICoreWebView2CreateCoreWebView2ControllerCompletedHandler};
-
-static constexpr auto environment_completed =
-    cast_info_t<ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler>{
-        IID_ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler};
-
-static constexpr auto message_received =
-    cast_info_t<ICoreWebView2WebMessageReceivedEventHandler>{
-        IID_ICoreWebView2WebMessageReceivedEventHandler};
-
-static constexpr auto permission_requested =
-    cast_info_t<ICoreWebView2PermissionRequestedEventHandler>{
-        IID_ICoreWebView2PermissionRequestedEventHandler};
-
-static constexpr auto add_script_to_execute_on_document_created_completed =
-    cast_info_t<
-        ICoreWebView2AddScriptToExecuteOnDocumentCreatedCompletedHandler>{
-        IID_ICoreWebView2AddScriptToExecuteOnDocumentCreatedCompletedHandler};
-} // namespace cast_info
-} // namespace mswebview2
 
 // Checks whether the specified IID equals the IID of the specified type and
 // if so casts the "this" pointer to T and returns it. Returns nullptr on
